added ability to save downloaded json to database
