+++ src/main/java/com/apzumi/apzumiapi/DemoApplication.java
@@ -1,9 +1,14 @@
 package com.apzumi.apzumiapi;
 
+import com.apzumi.apzumiapi.domain.Post;
+import com.apzumi.apzumiapi.service.PostService;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.client.RestTemplate;
+
+import java.util.List;
 
 @SpringBootApplication
 public class DemoApplication {
@@ -15,6 +20,15 @@ public class DemoApplication {
     @Bean
     public RestTemplate restTemplate() {
         return new RestTemplate();
+    }
+
+    @Bean
+    CommandLineRunner runner(PostService postService) {
+        return args -> {
+            List<Post> posts = postService.downloadPosts();
+            postService.save(posts);
+            System.out.println("Posts Saved!");
+        };
     }
 
 }
