--- src/main/java/com/apzumi/apzumiapi/PostRepository.java
@@ -1,9 +1,0 @@
-package com.apzumi.apzumiapi;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends CrudRepository<Posts, Long> {
-
-}
