+++ src/main/java/com/apzumi/apzumiapi/domain/Post.java
@@ -1,0 +1,25 @@
+package com.apzumi.apzumiapi.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Data
+@AllArgsConstructor
+@Entity
+public class Post {
+
+    public int userId;
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    public Long id;
+    public String title;
+    public String body;
+
+    public Post() {
+    }
+}
