+++ src/main/java/com/apzumi/apzumiapi/repository/PostRepository.java
@@ -1,0 +1,10 @@
+package com.apzumi.apzumiapi.repository;
+
+import com.apzumi.apzumiapi.domain.Post;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepository extends CrudRepository<Post, Long> {
+
+}
