+++ src/main/resources/application.properties
@@ -1,2 +1,2 @@
-spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.url=jdbc:h2:mem:posts
 spring.h2.console.enabled=true
